Removed utils sub-project
