+++ elastiknn-lucene/src/main/java/com/klibisz/elastiknn/models/KthGreatest.java
@@ -1,38 +1,6 @@
-package com.klibisz.elastiknn.utils;
+package com.klibisz.elastiknn.models;
 
-/**
- * Java implementations of some hot spots.
- */
-public class ArrayUtils {
-
-    /**
-     * Compute the number of intersecting (i.e. identical) elements between two int arrays.
-     * IMPORTANT: Assumes the given arrays are already sorted in ascending order and _does not_ check if this is true.
-     * If the given arrays are not sorted, the answer will be wrong.
-     * This is implemented in Java because for some reason Scala will Box the ints in some cases which is unnecessary
-     * and far slower.
-     * @param xs
-     * @param ys
-     * @return The number of identical elements in the two arrays. For example {1,2,3}, {2,3,4} would return 2.
-     */
-    public static int sortedIntersectionCount(final int [] xs, final int [] ys) {
-        int n = 0;
-        int xi = 0;
-        int yi = 0;
-        while (xi < xs.length && yi < ys.length) {
-            int x = xs[xi];
-            int y = ys[yi];
-            if (x < y) xi += 1;
-            else if (x > y) yi += 1;
-            else {
-                n += 1;
-                xi += 1;
-                yi += 1;
-            }
-        }
-        return n;
-    }
-
+public class KthGreatest {
     /**
      * Find the kth greatest value in the given array of shorts in O(N) time and space.
      * Works by creating a histogram of the array values and traversing the histogram in reverse order.
